Make UI for products fragment, make adapter, show data. Add logging interceptor. Change Long to Doublt in ProductItemDto. Make navigate from ProductsFragment to AddProductFragment.

+++ app/src/main/java/com/getswipe/android/data/dto/ProductsItemDto.kt
@@ -7,21 +7,21 @@ data class ProductsItemDto(
     @SerializedName("image")
     val image: String?,
     @SerializedName("price")
-    val price: Long?,
+    val price: Double?,
     @SerializedName("product_name")
     val productName: String?,
     @SerializedName("product_type")
     val productType: String?,
     @SerializedName("tax")
-    val tax: Long?,
+    val tax: Double?,
 )
 
 fun ProductsItemDto.toProductModel(): ProductModel {
     return ProductModel(
         image = image,
-        price = price ?: 0L,
+        price = price ?: 0.0,
         productName = productName ?: "",
         productType = productType ?: "",
-        tax = tax ?: 0L
+        tax = tax ?: 0.0,
     )
 }
