+++ app/src/main/java/com/getswipe/android/domain/model/ProductModel.kt
@@ -2,8 +2,8 @@ package com.getswipe.android.domain.model
 
 data class ProductModel(
     val image: String?,
-    val price: Long,
+    val price: Double,
     val productName: String,
     val productType: String,
-    val tax: Long,
+    val tax: Double,
 )
